hackerRank #Search, TripleSum3Arrays,  fix TestCase4: 86308 64827 78035

+++ src/main/java/com/sjia/hackerRank/TripleSum3Arrays.java
@@ -38,7 +38,7 @@ public class TripleSum3Arrays {
             } else {
                 tmp2 = -(tmp2 + 1);
             }
-            res += tmp1 * tmp2;
+            res += (long)tmp1 * tmp2;
         }
